requirements.txt
asyncpg>=0.29.0
crawl4ai>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0
aiohttp>=3.8.0
pandas>=2.0.0
psutil>=5.9.0
pytest>=7.0.0
pytest-asyncio>=0.21.0 